build(deps): bump pydantic from 2.0.3 to 2.4.0

+++ src/pixelator/report/models/graph.py
@@ -75,7 +75,7 @@ class GraphSampleReport(SampleReport):
         return self.a_pixel_count / self.b_pixel_count
 
     @pydantic.computed_field(  # type: ignore
-        return_type=float,
+        return_type=int,
         description="The total number of unique pixels in the graph.",
     )
     @property
